rename OSHDBTimeable to OSHDBTemporal
