+++ oshdb-tool/etl/src/main/java/org/heigit/ohsome/oshdb/tool/importer/transform/oshdb/OSMEntity.java
@@ -1,21 +1,26 @@
 package org.heigit.ohsome.oshdb.tool.importer.transform.oshdb;
 
 import java.util.Arrays;
-import org.heigit.ohsome.oshdb.OSHDBTimeable;
+import org.heigit.ohsome.oshdb.OSHDBTemporal;
 
-public interface OSMEntity extends OSHDBTimeable {
+public interface OSMEntity extends OSHDBTemporal {
 
-  public long getId();
-  public boolean isVisible();
-  
-  public int getVersion();
-  public long getChangeset();
-  public int getUserId();
-  public int[] getTags();
- 
+  long getId();
+
+  boolean isVisible();
+
+  int getVersion();
+
+  long getChangeset();
+
+  int getUserId();
+
+  int[] getTags();
+
 
   public default String asString() {
     return String.format("ID:%d V:%d TS:%d CS:%d VIS:%s UID:%d TAGS:%S", getId(), getVersion(),
-        getTimestamp().getEpochSecond(), getChangeset(), isVisible(), getUserId(), Arrays.toString(getTags()));
+        getTimestamp().getEpochSecond(), getChangeset(), isVisible(), getUserId(),
+        Arrays.toString(getTags()));
   }
 }
